apps/agentic-guardrails/services/gibberish-detection/requirements.txt
akto-llm-guard==0.3.17
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
numpy<2
torch>=2.4.0
optimum[onnxruntime]>=1.16.0